pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hispanie-event-card"
version = "0.1.0"
description = "Event card generator with bilingual support and weather integration"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Daniel Ducuara", email = "danielducuara@example.com" }]
keywords = ["event", "card", "generator", "bilingual", "weather"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "emoji==1.7.0",
    "pillow==11.3.0",
    "pilmoji==2.0.3",
    "requests==2.32.5",
]

[project.optional-dependencies]
dev = ["pre-commit>=4.0.1", "pytest"]

[project.urls]
Homepage = "https://github.com/danielducuara/hispanie-event-card"
Repository = "https://github.com/danielducuara/hispanie-event-card"
Issues = "https://github.com/danielducuara/hispanie-event-card/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["hipanie_event_card*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
